Add Illegal Spacing Check for Name and Address Parsing
